feat: update offer model to change timeLimit type from Float to Int
feat: integrate expireTimer middleware for trade expiration handling

feat: modify createOfferController to handle timeLimit in seconds

feat: implement cron job to expire trades based on time limits

fix: update getCreateTradeDetails to handle timeLimit correctly

feat: add socket service for managing Socket.IO instance

feat: enhance chat handler to manage trade timer updates and expirations

feat: create timer handler for socket events related to trade timers

refactor: streamline trade socket handler to manage trade state updates

style: improve trade status display in frontend with new styles

feat: add remaining time formatting utility for better UX

fix: update trade-related types and hooks to reflect new state management

style: enhance global styles for trade status indicators

fix: adjust trade page to display remaining time and status correctly

+++ backend/prisma/migrations/20250526093227_/migration.sql
@@ -142,7 +142,7 @@ CREATE TABLE "offers" (
     "listAt" DOUBLE PRECISION NOT NULL,
     "limitMin" DOUBLE PRECISION NOT NULL,
     "limitMax" DOUBLE PRECISION NOT NULL,
-    "timeLimit" DOUBLE PRECISION NOT NULL,
+    "timeLimit" INTEGER NOT NULL,
     "tags" TEXT[],
     "label" TEXT NOT NULL,
     "terms" TEXT NOT NULL,
